requirements.txt
# 敦煌100MW光热电站定日镜场优化项目依赖
# Scientific Data 数据集生成工具

# 核心科学计算库
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 可视化库
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# 地理空间数据处理
geopandas>=0.9.0
rasterio>=1.2.0
shapely>=1.7.0
fiona>=1.8.0
pyproj>=3.1.0

# 图像处理
Pillow>=8.0.0
opencv-python>=4.5.0

# 优化算法
scikit-learn>=1.0.0
deap>=1.3.0

# 数据格式支持
openpyxl>=3.0.0
xlrd>=2.0.0
h5py>=3.1.0
netcdf4>=1.5.0

# 网络请求
requests>=2.25.0
urllib3>=1.26.0

# 进度条和日志
tqdm>=4.60.0
loguru>=0.5.0

# 配置文件处理
pyyaml>=5.4.0
toml>=0.10.0

# 数据验证
jsonschema>=3.2.0
cerberus>=1.3.0

# 并行计算
joblib>=1.0.0
multiprocessing-logging>=0.3.0

# 统计分析
statsmodels>=0.12.0

# 时间处理
python-dateutil>=2.8.0
pytz>=2021.1

# 开发和测试工具
pytest>=6.0.0
pytest-cov>=2.10.0
black>=21.0.0
flake8>=3.8.0

# Jupyter支持 (可选)
jupyter>=1.0.0
ipykernel>=6.0.0

# 文档生成 (可选)
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# 数据集发布工具
zenodo-client>=0.3.0